fix(components): [table]
toggleRowStatus fix
